Comment System Done

+++ src/pages/Main/Winwin/Wdetail.js
@@ -1,57 +1,84 @@
 import React, { useCallback, useEffect, useState } from "react";
 import ReactQuill from "react-quill";
 import { getWinOneBoard } from "../../../apiService/winBoardApiService";
-import { deleteComment, fetchComments, insertComment } from "../../../apiService/CommentApiService";
+import { deleteComment, fetchComments, insertComment, updateComment } from "../../../apiService/CommentApiService";
 import { useSelector } from "react-redux";
 import "./Wdetail.scss";
 import ParentsComments from "../components/Comment/ParentsComments";
 
 const Wdetail = ({ match }) => {
+  //게시글 상세정보
   const [wDetails, setWDetails] = useState({
     title: "",
     content: "",
+    user: {
+      id: 0,
+    },
   });
+
+  //대댓글을 등록중인지 확인하는 state
+  const [isReplying, setIsReplying] = useState({
+    isReplying: false,
+    reply_commentId: 0,
+  });
+
+  //전체 댓글관리
   const [comments, setComments] = useState();
+
+  //root댓글 input
   const [inputReply, setInputReply] = useState({
     content: "",
   });
+
+  //접속중인 사람의 정보
   const { userData } = useSelector((state) => state.loginReducer);
+
+  //게시글 상세정보 및 댓글 가져오기
   useEffect(() => {
     getWinOneBoard(match.params.board_id)
       .then((res) => {
         setWDetails(res.data);
       })
       .catch((e) => {
-        alert(e.response.message);
+        alert(e.response.data.message);
       });
     fetchComments(match.params.board_id)
       .then((res) => {
         setComments(res.data);
       })
       .catch((e) => {
-        console.log(e);
+        alert(e.response.data.message);
       });
   }, [match.params.board_id]);
+
+  //댓글 삭제
   const deleteReply = useCallback(
     (commentId) => {
-      deleteComment(commentId)
-        .then(() => {
-          fetchComments(match.params.board_id)
-            .then((res) => {
-              setComments(res.data);
-            })
-            .catch((e) => {
-              console.log(e.response);
-            });
-        })
-        .catch((e) => {
-          console.log(e.response);
-        });
+      if (window.confirm("댓글을 삭제하시겠습니까?")) {
+        deleteComment(commentId)
+          .then(() => {
+            fetchComments(match.params.board_id)
+              .then((res) => {
+                setComments(res.data);
+              })
+              .catch((e) => {
+                alert(e.response.data.message);
+              });
+          })
+          .catch((e) => {
+            alert(e.response.data.message);
+          });
+      }
     },
     [match.params.board_id]
   );
 
+  //root댓글입력 저장
   const insertReply = useCallback(() => {
+    if (inputReply.content === "") {
+      alert("내용을 입력해 주세요");
+      return;
+    }
     const insertData = {
       ...inputReply,
       userId: userData.id,
@@ -66,14 +93,15 @@ const Wdetail = ({ match }) => {
             setInputReply({ content: "" });
           })
           .catch((e) => {
-            console.log(e.response);
+            alert(e.response.data.message);
           });
       })
       .catch((e) => {
-        console.log(e.response);
+        alert(e.response.data.message);
       });
   }, [match.params.board_id, userData, inputReply]);
 
+  //Root 댓글 핸들러
   const replyInputHandler = useCallback(
     (e) => {
       setInputReply({
@@ -83,6 +111,80 @@ const Wdetail = ({ match }) => {
     },
     [inputReply]
   );
+
+  // 대댓글 창열기
+  const reReplyOpen = useCallback(
+    (commnetId) => {
+      setIsReplying({
+        ...isReplying,
+        replying: true,
+        reply_commentId: commnetId,
+      });
+    },
+    [isReplying]
+  );
+
+  // 대댓글 입력 저장
+  const reReplyInsert = useCallback(
+    (reReplyData) => {
+      if (reReplyData.content === "") {
+        alert("내용을 입력해 주세요");
+        return;
+      }
+      const insertData = {
+        ...reReplyData,
+        userId: userData.id,
+        boardId: match.params.board_id,
+      };
+      insertComment(insertData)
+        .then(() => {
+          fetchComments(match.params.board_id)
+            .then((res) => {
+              setComments(res.data);
+            })
+            .catch((e) => {
+              alert(e.response.data.message);
+            });
+        })
+        .catch((e) => {
+          alert(e.response.data.message);
+        });
+    },
+    [match.params.board_id, userData]
+  );
+
+  //댓글 수정중인지 확인
+  const [isModifying, setIsModifying] = useState({
+    isModifying: false,
+    modify_commentId: 0,
+  });
+  //댓글 수정 저장
+  const modifyComment = useCallback(
+    (modifyData) => {
+      if (modifyData.content === "") {
+        alert("내용을 입력해 주세요");
+        return;
+      }
+      const modiContent = {
+        content: modifyData.content,
+      };
+      updateComment(modifyData.commentId, modiContent)
+        .then(() => {
+          fetchComments(match.params.board_id)
+            .then((res) => {
+              setComments(res.data);
+            })
+            .catch((e) => {
+              alert(e.response.data.message);
+            });
+        })
+        .catch((e) => {
+          alert(e.response.data.message);
+        });
+    },
+    [match.params.board_id]
+  );
+
   return (
     <div>
       <div className='detail-content'>
@@ -94,17 +196,25 @@ const Wdetail = ({ match }) => {
           <ul>
             {comments &&
               comments.map((comment, index) => (
-                <ParentsComments
-                  writer={wDetails.user.id}
-                  userData={userData}
-                  comment={comment}
-                  key={index}
-                  deleteReply={deleteReply}
-                />
+                <React.Fragment key={index}>
+                  <ParentsComments
+                    writer={wDetails.user.id}
+                    userData={userData}
+                    comment={comment}
+                    deleteReply={deleteReply}
+                    reReplyOpen={reReplyOpen}
+                    isReplying={isReplying}
+                    setIsReplying={setIsReplying}
+                    reReplyInsert={reReplyInsert}
+                    isModifying={isModifying}
+                    setIsModifying={setIsModifying}
+                    modifyComment={modifyComment}
+                  />
+                </React.Fragment>
               ))}
           </ul>
 
-          {/* 댓글 다는 곳 */}
+          {/* root댓글 다는 곳 */}
           <textarea name='content' value={inputReply.content} className='textarea' onChange={replyInputHandler} />
           <button onClick={insertReply}>댓글 등록</button>
         </div>
